Update header files

+++ include/Lt-Core.hh
@@ -44,7 +44,7 @@ typedef struct{
     XEvent Xe;
 } D_Lt_Core;
 
-void Lt_Core(Display *Dsp, int _CONS);
+void Lt_Core(Display *Dsp, short int _CONS);
 void Lt_Read(const char *_FILE[]);
 void Lt_Write(const char *_FILE[]);
 void Lt_LineCounter();
